feat(profile): wrap profile page with Suspense and ProfileResolver
Phase 3: /social/profile/[username] を Suspense + Skeleton 構造に書き換え。

- ProfileResolver.tsx (新規 Server Component) に UUID resolve + redirect ロジックを
  分離し、Suspense 境界の内側で動かす
- SocialProfileSkeleton.tsx (新規) は ProfileCardSkeleton/StatsCardSkeleton/
  ProfileTabBarSkeleton/SocialPostCardSkeleton×3 を組み合わせた合成 Skeleton
- page.tsx は SocialLayout の DOM (.layout > .contentWrapper > .main) を inline で
  再現し、その内側に Suspense + ProfileResolver を配置
  (SocialLayout コンポーネント本体を直接 wrap すると、Next.js 16 + Turbopack の
   SSR module 評価で TabNavigation → useAuth → createContext のチェーンが評価
   されて build に失敗するため、CSS Modules だけ流用して回避)
- SocialProfileSkeleton 内では SocialHeader を呼ばず inline の header skeleton で
  代替 (SocialHeader → Button (forwardRef) も同様の SSR module 評価エラーを起こす
   ため)

cacheComponents は未有効のままなので真の PPR 効果は出ないが、Suspense ストリーミング
により SSR 完了待ちが解消され Skeleton が即時表示される。
